Rename instance-searching tools
