+++ source/node/tools/vaultSearch.js
